requirements.txt
pyyaml
pylint
pydantic
